+++ .lintstagedrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  "*.{js,ts}": "npm run lint",
+  "*.{ts,js,json,md}": "npm run prettier:format",
+  "*": "npm run build",
+};