+++ src/utils/base64.ts
@@ -1,10 +1,7 @@
 export function escape(value: string): string {
-    return value
-        .replace(/=/g, '')
-        .replace(/\+/g, '-')
-        .replace(/\//g, '_')
+  return value.replace(/=/g, "").replace(/\+/g, "-").replace(/\//g, "_");
 }
 
 export function toBase64(value: Buffer): string {
-    return escape(value.toString('base64'))
+  return escape(value.toString("base64"));
 }