+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Minor Changes
 
 - Added automated Heartbeat messages
-  - new option `heartbeatIntervalMs` DEFAULT: 5 * 60 * 1000
+  - new option `heartbeatIntervalMs` DEFAULT: 5 _ 60 _ 1000
   - new events `ON_HEARTBEAT` - this is emited when socket recieves `ping` or `ack` messages
 
 ### Patch Changes